+++ framework/console/controllers/MigrateController.php
@@ -154,6 +154,7 @@ class MigrateController extends BaseMigrateController
             't' => 'migrationTable',
             'F' => 'templateFile',
             'P' => 'useTablePrefix',
+            'c' => 'compact',
         ]);
     }
 
@@ -184,7 +185,7 @@ class MigrateController extends BaseMigrateController
     protected function createMigration($class)
     {
         $this->includeMigrationFile($class);
-        return new $class(['db' => $this->db]);
+        return new $class(['db' => $this->db, 'compact' => $this->compact]);
     }
 
     /**